+++ R/flowcell.R
@@ -160,3 +160,33 @@ Flowcell <- R6::R6Class(
 
   )
 )
+
+
+#' Prepare a `flowcell` object from provided class
+#'
+#' This method is intended to streamline the usage of floundeR in more coherent
+#' workflows through the enabling of more magrittr-like piped commands. This
+#' method will consume floundeR specified classes and return a flowcell object
+#' where possible.
+#'
+#' @param r6_object is the floundeR R6 object to extract flowcell from
+#'
+#' @return Flounder flowcell object
+#' 
+#' @importFrom magrittr %>%
+#' @export %>%
+#'
+#' @examples
+#' SequencingSummary$new(flnDr("sequencing_summary.txt.bz2")) %>% to_flowcell()
+#'
+#' @export
+to_flowcell = function(r6_object) {
+  if (class(r6_object)[1] == "SequencingSummary") {
+    return(r6_object$flowcell)
+  } else {
+    stop("Unable to prepare flowcell object from provided input")
+  }
+  
+}
+
+