+++ tests/testthat/test-sequencing_summary.R
@@ -11,8 +11,6 @@ test_that("sequencing_summary", {
     # and check that the specified extant path is actually a file (not dir)
     expect_error(SequencingSummary$new(".."))
     expect_error(seqsum <- SequencingSummary$new(sequencing_summary), NA)
-    expect_equal(class(seqsum$flowcell)[1], "Flowcell")
-    # please note that Flowcell testing is continued in the Flowcell unit tests
 })
 
 
@@ -40,6 +38,25 @@ test_that("sequencing_summary + barcode", {
         seqsumBC <- SequencingSummary$new(
             sequencing_summary_file=sequencing_summary,
             barcoding_summary_file=barcodes_summary), NA)
-    expect_equal(class(seqsumBC$flowcell)[1], "Flowcell")
+})
+
+
+
+context("Flowcell object creation")
+test_that("Flowcell object creation", {
+    
+    sequencing_summary <- flnDr("sequencing_summary.txt.bz2")
+    barcodes_summary <- flnDr("barcoding_summary.txt.bz2")
+
+    seqsum <- SequencingSummary$new(
+        sequencing_summary_file=sequencing_summary,
+        barcoding_summary_file=barcodes_summary)
+    
+    expect_equal(class(seqsum$flowcell)[1], "Flowcell")
+    
+    flowcell <- seqsum %>% to_flowcell()
+    expect_equal(class(flowcell)[1], "Flowcell")
+    
+
     # please note that Flowcell testing is continued in the Flowcell unit tests
 })