+++ NEWS.md
@@ -1,3 +1,5 @@
+# floundeR (development version)
+
 # floundeR 0.0.3
 
 * pkgdown documentation for the github pages