+++ NEWS.md
@@ -1,5 +1,7 @@
 # floundeR (development version)
 
+* starting to ensure that methods are also available through documented
+  magrittr %>% pipes for legibility of code
 * robust Angenieux decorations and demonstration of more comprehensive plots
 * integrating Angenieux with ggsave for cleaner Rmarkdown usage